Reorganized objects into seperate files
