+++ openFrameworks/apps/Ray Tracing Renderer/Ray Tracing Renderer.vcxproj
@@ -194,13 +194,16 @@
   <ItemGroup>
     <ClCompile Include="src\main.cpp" />
     <ClCompile Include="src\ofApp.cpp" />
+    <ClCompile Include="src\rtGraphics\Objects\meshObject.cpp" />
+    <ClCompile Include="src\rtGraphics\Objects\sphereObject.cpp" />
     <ClCompile Include="src\rtGraphics\rtColorf.cpp" />
-    <ClCompile Include="src\rtGraphics\rtObject.cpp" />
     <ClCompile Include="src\rtGraphics\rtRenderer.cpp" />
     <ClCompile Include="src\rtGraphics\rtVec3f.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\ofApp.h" />
+    <ClInclude Include="src\rtGraphics\Objects\meshObject.h" />
+    <ClInclude Include="src\rtGraphics\Objects\sphereObject.h" />
     <ClInclude Include="src\rtGraphics\rtCam.h" />
     <ClInclude Include="src\rtGraphics\rtColorf.h" />
     <ClInclude Include="src\rtGraphics\rtRayHit.h" />
