+++ openFrameworks/apps/Ray Tracing Renderer/Ray Tracing Renderer.vcxproj.filters
@@ -10,6 +10,9 @@
     <Filter Include="src\rtGraphics">
       <UniqueIdentifier>{98fd35b3-6c03-446f-80bd-7ca0480b92f0}</UniqueIdentifier>
     </Filter>
+    <Filter Include="src\rtGraphics\Objects">
+      <UniqueIdentifier>{2cb30c58-4576-4d1d-8b72-79fbba8af173}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\main.cpp">
@@ -24,11 +27,14 @@
     <ClCompile Include="src\rtGraphics\rtColorf.cpp">
       <Filter>src\rtGraphics</Filter>
     </ClCompile>
-    <ClCompile Include="src\rtGraphics\rtObject.cpp">
-      <Filter>src\rtGraphics</Filter>
-    </ClCompile>
     <ClCompile Include="src\rtGraphics\rtRenderer.cpp">
       <Filter>src\rtGraphics</Filter>
+    </ClCompile>
+    <ClCompile Include="src\rtGraphics\Objects\sphereObject.cpp">
+      <Filter>src\rtGraphics\Objects</Filter>
+    </ClCompile>
+    <ClCompile Include="src\rtGraphics\Objects\meshObject.cpp">
+      <Filter>src\rtGraphics\Objects</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -70,6 +76,12 @@
     </ClInclude>
     <ClInclude Include="src\rtGraphics\rtRayHit.h">
       <Filter>src\rtGraphics</Filter>
+    </ClInclude>
+    <ClInclude Include="src\rtGraphics\Objects\sphereObject.h">
+      <Filter>src\rtGraphics\Objects</Filter>
+    </ClInclude>
+    <ClInclude Include="src\rtGraphics\Objects\meshObject.h">
+      <Filter>src\rtGraphics\Objects</Filter>
     </ClInclude>
   </ItemGroup>
 </Project>
