+++ openFrameworks/apps/Ray Tracing Renderer/src/ofApp.cpp
@@ -10,12 +10,12 @@ void ofApp::setup()
 
 	///Create a red, matte sphere and add it to the scene
 	rtMat shinyRed(rtColorf(0.2f, 0.0f, 0.0f), rtColorf::red, rtColorf::white, 200.0f);
-	rtObject* redSphere = new rtSphere(rtVec3f(25.0f, -25.0f, 0.0f), 20.0f, shinyRed);
+	rtObject* redSphere = new sphereObject(rtVec3f(25.0f, -25.0f, 0.0f), 20.0f, shinyRed);
 	demoScene->addObject("red sphere", redSphere);
 
 	///Create a blue, reflective sphere and add it to the scene
 	rtMat reflectiveBlue(rtColorf(0.0f, 0.0f, 0.2f), rtColorf::blue, rtColorf::white, 500.0f, 0.3f);
-	rtObject* blueSphere = new rtSphere(rtVec3f(-25.0f, -25.0f, 0.0f), 20.0f, reflectiveBlue);
+	rtObject* blueSphere = new sphereObject(rtVec3f(-25.0f, -25.0f, 0.0f), 20.0f, reflectiveBlue);
 	demoScene->addObject("blue sphere", blueSphere);
 	
 	///Create a mesh and add it to the scene
@@ -26,7 +26,7 @@ void ofApp::setup()
 	triangleMesh.addFace(0, 1, 2);
 
 	rtMat matteGreen(rtColorf(0.0f, 0.2f, 0.0f), rtColorf::green, rtColorf(0.6f), 20.0f);
-	rtObject* triangle = new rtMeshObject(triangleMesh, matteGreen);
+	rtObject* triangle = new meshObject(triangleMesh, matteGreen);
 	demoScene->addObject("triangle", triangle);
 
 	///Create a box surrounding the scene
@@ -48,7 +48,7 @@ void ofApp::setup()
 	boxMesh.addFace(4, 0, 1); boxMesh.addFace(1, 5, 4);		//Bottom Wall
 
 	rtMat matteWhite(rtColorf(0.2f), rtColorf::white, rtColorf(0.4f), 20.0f);
-	rtObject* box = new rtMeshObject(boxMesh, matteWhite);
+	rtObject* box = new meshObject(boxMesh, matteWhite);
 	demoScene->addObject("box", box);
 
 	///Create a light and add it to the scene
