+++ openFrameworks/apps/Ray Tracing Renderer/src/rtGraphics/rtMain.h
@@ -2,3 +2,5 @@
 
 #include "rtCam.h"
 #include "rtScene.h"
+#include "Objects/sphereObject.h"
+#include "Objects/meshObject.h"
